feature(Artisan): Move out and parametrize crafting table
